IC-2927

+++ i21Database/dbo/Stored Procedures/uspICPostCostAdjustmentOnLotCosting.sql
@@ -148,7 +148,7 @@ DECLARE @InvTranId AS INT
 		,@InvTranSubLocationId AS INT
 		,@InvTranStorageLocationId AS INT 
 		,@InvTranCurrencyId AS INT
-		,@InvTranExchangeRate AS INT
+		,@InvTranExchangeRate AS NUMERIC(38,20)
 		,@InvTranStringTransactionId AS NVARCHAR(40)
 		,@InvTranIntTransactionId AS INT
 		,@InvTranTypeName AS NVARCHAR(200)
